fixing keycode and key speed on drop down
bug: 2player drop doesn't work, freezes the game once the block is
placed which casuse 1player to crash also

+++ main.js
@@ -61,27 +61,38 @@ const playerElements = document.querySelectorAll('.player');
     tetri.push(tetris);
 });
 
-
-
-// Using Keycode instead of key
-document.addEventListener('keydown', event => {
+const keyListener = (event) => {
   [
     [65, 68, 81, 69, 83], //left = a, right = d , rotate q&e. drop s = Player One
     [72, 75, 89, 73, 74], //left = h, right = k , rotate y&i. drop j = Player Two
   ].forEach((key, index) => {
       const player = tetri[index].player;
-      if (event.keyCode === key[0]) {
-        player.move(-1);
-      } else if (event.keyCode === key[1]) {
-        player.move(1);
-      } else if (event.keyCode === key[2]) {
-        player.rotate(-1);
-      } else if (event.keyCode === key[3]) {
-        player.rotate();
-      } else if (event.keyCode === key[4]) {
-        player.drop();
+      if (event.type === 'keydown') {
+        if (event.keyCode === key[0]) {
+          player.move(-1);
+        } else if (event.keyCode === key[1]) {
+          player.move(1);
+        } else if (event.keyCode === key[2]) {
+          player.rotate(-1);
+        } else if (event.keyCode === key[3]) {
+          player.rotate();
+        }
       }
-    }); 
-});
+
+      if (event.keyCode === key[4]) {
+          if (event.type === 'keydown') {
+              player.drop();
+              player.dropInterval = player.DROP_FAST;
+            } else {
+                player.dropInterval = player.DROP_SLOW;
+            }
+        
+      }
+  }); 
+};
+
+
+document.addEventListener('keydown', keyListener);
+document.addEventListener('keyup', keyListener);
 
 updateScore();
